requirements.txt
flask
flask-session
werkzeug
mysql-connector-python
pillow  
numpy  
opencv-python  
